refactor: placeholder 이미지 webp 변환(#596)
