+++ src/components/commons/card/ChatProductCard.tsx
@@ -1,4 +1,4 @@
-import PlaceholderImage from '@assets/images/placeholder.png'
+import PlaceholderImage from '@assets/images/placeholder.webp'
 import { formatPrice } from '@src/utils/formatPrice'
 
 interface ChatProductCardProps {
