+++ src/pages/product-detail/components/SubImages.tsx
@@ -1,4 +1,4 @@
-import PlaceholderImage from '@assets/images/placeholder.png'
+import PlaceholderImage from '@assets/images/placeholder.webp'
 
 interface SubImagesProps {
   mainImageUrl: string
